[Photo] Create Photo class with mostly cut+pasted code from PhotoPreviewView
Should have done this a while ago to be honest.
